requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Dashboard
streamlit>=1.28.0

# Model Interpretability
shap>=0.42.0

# Network Capture (optional)
scapy>=2.5.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Imbalanced Learning
imbalanced-learn>=0.11.0
